+++ sctplib/programs/echo_tool.c
@@ -1,5 +1,5 @@
 /*
- *  $Id: echo_tool.c,v 1.7 2003/11/21 14:40:45 tuexen Exp $
+ *  $Id: echo_tool.c,v 1.8 2004/11/17 20:56:06 tuexen Exp $
  *
  * SCTP implementation according to RFC 2960.
  * Copyright (C) 2000 by Siemens AG, Munich, Germany.
@@ -287,7 +287,7 @@ void checkArgs(void)
         exit(-1);
 }
 
-void dataArriveNotif(unsigned int assocID, unsigned int streamID, unsigned int len,
+void dataArriveNotif(unsigned int assocID, unsigned short streamID, unsigned int len,
                      unsigned short streamSN,unsigned int TSN, unsigned int protoID,
                      unsigned int unordered, void* ulpDataPtr)
 {
