+++ sctplib/programs/monitor.c
@@ -523,7 +523,7 @@ void* communicationUpNotif(unsigned int assocID, int status, unsigned int noOfDe
   return NULL;
 }
 
-void dataArriveNotif(unsigned int assocID, unsigned int streamID, unsigned int len,
+void dataArriveNotif(unsigned int assocID, unsigned short streamID, unsigned int len,
                      unsigned short streamSN,unsigned int TSN, unsigned int protoID,
                      unsigned int unordered, void* ulpDataPtr)
 {
